Delete settings on stock-portfolio deletion. Use transaction

+++ src/server/graphql/nexus/mutations/stock-portfolio.mutations/delete-one-stock-portfolio.mutation.ts
@@ -1,5 +1,6 @@
 import { PrismaUtils } from "@/server/utils";
 import { arg, extendType } from "@nexus/schema";
+import { last } from "lodash";
 
 export const deleteOneStockPortfolio = extendType({
 	type: "Mutation",
@@ -32,10 +33,23 @@ export const deleteOneStockPortfolio = extendType({
 
 				return isOwnedByUser;
 			},
-			resolve: (parent, args, { prisma }) => {
+			resolve: async (parent, args, { prisma }) => {
 				const { where } = PrismaUtils.castInputs(args);
 
-				return prisma.stockPortfolio.delete({ where });
+				const stockPortfolio = await prisma.stockPortfolio.findOne({ where });
+
+				if (!stockPortfolio) {
+					return null;
+				}
+
+				await prisma.transaction([
+					prisma.stockPortfolioSettings.delete({
+						where: { stockPortfolioId: stockPortfolio.id }
+					}),
+					prisma.stockPortfolio.delete({ where })
+				]);
+
+				return stockPortfolio;
 			}
 		});
 	}
